separing_filtering_scrapping
