+++ .gitignore
@@ -3,6 +3,7 @@ data_results/
 *.ipynb
 *.png
 *.csv
+*.json
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
